requirements.txt
# requirements.txt
pillow
numpy
matplotlib
